move helper codes one level
